ORCH-0965 fix: break require-cycle (I-PROPOSED-K) by extracting upcomingKeys to its own module
